+++ frontend/src/pages/organizations/index.tsx
@@ -1,32 +1,16 @@
-import { gql } from "@apollo/client";
 import { createGetServerSideProps } from "core/helpers/page";
 import { NextPageWithLayout } from "core/helpers/types";
 import OrganizationsLayout from "./OrganizationsLayout";
 import { useTranslation } from "next-i18next";
 import Page from "core/components/Page";
+import {
+  OrganizationsDocument,
+  OrganizationsQuery,
+} from "organizations/graphql/queries.generated";
 
 type Props = {
-  organizations: {
-    id: string;
-    name: string;
-    workspaces: { items: { slug: string; name: string }[] };
-  }[];
+  organizations: OrganizationsQuery["organizations"];
 };
-
-const ORGANIZATIONS_QUERY = gql`
-  query Organizations {
-    organizations {
-      id
-      name
-      workspaces {
-        items {
-          slug
-          name
-        }
-      }
-    }
-  }
-`;
 
 const OrganizationsPage: NextPageWithLayout<Props> = ({ organizations }) => {
   const { t } = useTranslation();
@@ -44,7 +28,7 @@ export const getServerSideProps = createGetServerSideProps({
   async getServerSideProps(ctx, client) {
     await OrganizationsLayout.prefetch(ctx, client);
     const { data } = await client.query({
-      query: ORGANIZATIONS_QUERY,
+      query: OrganizationsDocument,
     });
 
     return {
